+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 5.6 (planned)
 
+* PR #73 optimized thread hook algorithm
 * PR #72 QaseStepListener javadoc fix
 
 ## 5.5 (released 25.11.2022)