+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## 5.6 (planned)
 
+* PR #79 added missed com.plugatar.xteps.base.hook package-info
 * PR #78 added step objects public ctor without step name
+* PR #77 Bump qase-api from 3.0.0 to 3.0.1
 * PR #74 added threadHookPriority option
 * PR #73 optimized thread hook algorithm
 * PR #72 QaseStepListener javadoc fix